requirements.txt
aniso8601==10.0.1
attrs==25.4.0
blinker==1.9.0
cffi==2.0.0
colorama==0.4.6
cryptography==46.0.3
distlib==0.4.0
et_xmlfile==2.0.0
Flask==3.1.2
flask-cors==6.0.1
Flask-Login==0.6.3
flask-restx==1.3.2
greenlet==3.2.4
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
mysql-connector==2.2.9
mysql-connector-python==9.4.0
numpy==2.0.2
openpyxl==3.1.5
pandas==2.3.3
platformdirs==4.4.0
pycparser==2.23
PyJWT==2.10.1
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
referencing==0.36.2
rpds-py==0.27.1
six==1.17.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
tzdata==2025.2
virtualenv==20.35.4
Werkzeug==3.1.3
